requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
autopep8==2.0.2
bcrypt==4.0.1
bidict==0.22.1
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
dnspython==2.3.0
fastapi==0.100.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
idna==3.4
kombu==5.2.4
marshmallow==3.19.0
motor==3.1.2
packaging==23.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.0.3
pydantic-settings==2.0.2
pydantic_core==2.3.0
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
python-engineio==4.5.1
python-multipart==0.0.6
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
sentry-sdk==1.30.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.21.1
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.1
